Separate Agent class into 3 classes: Agent, QA_Learning, Exploration_Learning.
Just because original Agent class got big. Theory is that this will be easier to maintain and alter.
Next I will be deeply optimizing this stuff.  
For instance, hash the lookup of the complete action version of an incomplete action in the sim
instead of traversing the list of actions.

git-svn-id: https://cmxt.sri.com/svn/bl/graphplanner/branches/chrisw@7554 db5f4523-d333-0410-b191-92aa9660ebcb